+++ db/migrate/20210908131121_change_support_category_name_to_title.rb
@@ -1,0 +1,6 @@
+class ChangeSupportCategoryNameToTitle < ActiveRecord::Migration[6.1]
+  def change
+    rename_column :support_categories, :name, :title
+    add_index :support_categories, :title, unique: true
+  end
+end