+++ db/migrate/20210908131247_add_default_support_category_timestamps.rb
@@ -1,0 +1,7 @@
+class AddDefaultSupportCategoryTimestamps < ActiveRecord::Migration[6.1]
+  def change
+    change_table :support_categories, bulk: true do |t|
+      t.timestamps default: -> { "CURRENT_TIMESTAMP" }
+    end
+  end
+end