+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog 1.0.0].
 
 ## [Unreleased]
 
+### Supported Unreleased
+
+- Added initial models for supported case management functions.
+- Added unique 6 digit ref for cases starting with 000001.
+
+### Specify Unreleased
+
 **Rich Data**
 - Integrate fully with DSI to gather names, email and organisation at authentication
 - Add env vars for DSI API `DFE_SIGN_IN_API_SECRET`, `DFE_SIGN_IN_API_ENDPOINT`